pyproject.toml
[project]
name = "2-1-game-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "fastapi>=0.121.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "mysql-connector-python>=9.5.0",
    "openai>=2.6.1",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
